MT1 User Migration - https://pm.zetainteractive.com:8443/browse/MT2-136

+++ app/Services/UserService.php
@@ -7,8 +7,6 @@
  */
 
 namespace App\Services;
-
-
 use Cartalyst\Sentinel\Sentinel;
 
 
@@ -83,7 +81,7 @@ class UserService
      */
     public function getAllUsersWithRolesNames()
     {
-        $users = $this->userRepo->select('id', 'email', 'first_name', 'last_name', 'last_login')
+        $users = $this->userRepo->select('id', 'email','username', 'first_name', 'last_name', 'last_login')
             ->with(['roles' => function ($query) {
                 $query->addSelect('name');
             }, 'activations'])->get();
@@ -101,7 +99,7 @@ class UserService
      */
     public function getUserWithRoles($id)
     {
-        $user = $this->userRepo->select('id', 'email', 'first_name', 'last_name', 'last_login')
+        $user = $this->userRepo->select('id', 'email','username', 'first_name', 'last_name', 'last_login')
             ->with(['roles' => function ($query) {
                 $query->addSelect('id');
             }])->get()->find($id);
@@ -111,6 +109,7 @@ class UserService
         return $user;
     }
 
+
     /**
      * @param $input
      * @param $roles
@@ -119,10 +118,12 @@ class UserService
     public function updateUserAndRoles($input, $roles, $id)
     {
         $user = $this->userRepo->findById($id);
-        $user->roles()->detach(); //remove all roles.
-        foreach ($roles as $role) {
-            $roleObject = $this->authObject->findRoleById($role);
-            $roleObject->users()->attach($user);
+        if($roles) {
+            $user->roles()->detach(); //remove all roles.
+            foreach ($roles as $role) {
+                $roleObject = $this->authObject->findRoleById($role);
+                $roleObject->users()->attach($user);
+            }
         }
         $this->userRepo->update($user, $input);
     }
